requirements.txt
streamlit==1.22.0
langchain==0.0.181
PyPDF2==3.0.1
python-dotenv==1.0.0
faiss-cpu==1.7.4
huggingface-hub==0.14.1sentence-transformers==2.2.2
